+++ commons/gridgo-utils/src/main/java/io/gridgo/utils/pojo/asm/setter/AsmSetterGenerator.java
@@ -10,8 +10,8 @@ import static org.objectweb.asm.Opcodes.V1_7;
 import org.objectweb.asm.ClassWriter;
 import org.objectweb.asm.MethodVisitor;
 
-import io.gridgo.utils.pojo.DynamicClassLoader;
 import io.gridgo.utils.pojo.PojoMethodSignature;
+import io.gridgo.utils.pojo.asm.DynamicClassLoader;
 import io.gridgo.utils.pojo.setter.PojoSetter;
 import io.gridgo.utils.pojo.setter.PojoSetterGenerator;
 
@@ -47,7 +47,7 @@ public class AsmSetterGenerator implements PojoSetterGenerator {
 
         cw.visitEnd();
 
-        Class<?> clz = DynamicClassLoader.getInstance().define(className, cw.toByteArray());
+        Class<?> clz = DynamicClassLoader.getInstance().loadByteCode(className, cw.toByteArray());
 
         try {
             return (PojoSetter) clz.getConstructor().newInstance();
